+++ lib/yaks/mapper/has_many.rb
@@ -1,7 +1,10 @@
 module Yaks
   class Mapper
     class HasMany < Association
+      include Util
+
       def map_resource(collection, context)
+        return NullResource.new(collection: true) if collection.nil?
         policy        = context.fetch(:policy)
         member_mapper = association_mapper(policy)
         context       = context.merge(member_mapper: member_mapper)
@@ -11,6 +14,10 @@ module Yaks
       def collection_mapper(collection, policy)
         return @collection_mapper unless @collection_mapper.equal? Undefined
         policy.derive_mapper_from_object(collection)
+      end
+
+      def singular_name
+        singularize(name.to_s)
       end
     end
   end