+++ spec/acceptance/acceptance_spec.rb
@@ -3,12 +3,12 @@ require 'spec_helper'
 require_relative './models'
 
 RSpec.shared_examples_for 'JSON output format' do |yaks, format, name|
-  let(:input)  { load_yaml_fixture name }
+  let(:input)  { load_yaml_fixture(name) }
   let(:output) { load_json_fixture "#{name}.#{format}" }
 
   subject { yaks.serialize(input) }
 
-  it { should eql output }
+  it { should deep_eql output }
 end
 
 RSpec.describe Yaks::Serializer::Hal do
@@ -22,8 +22,8 @@ RSpec.describe Yaks::Serializer::Hal do
     end
   end
 
-  include_examples 'JSON output format' , yaks_rel_template    , :hal      , 'confucius'
-  include_examples 'JSON output format' , yaks_policy_dsl      , :hal      , 'confucius'
+  include_examples 'JSON output format' , yaks_rel_template , :hal , 'confucius'
+  include_examples 'JSON output format' , yaks_policy_dsl   , :hal , 'confucius'
 end
 
 RSpec.describe Yaks::Serializer::JsonApi do